Save simple story to db (just an object with a name and a date)

+++ app/src/main/java/org/mbach/homeautomation/story/StoryDAO.java
@@ -8,23 +8,22 @@ package org.mbach.homeautomation.story;
  */
 public class StoryDAO {
 
-    private int id;
+    private long id;
     private String title;
 
     public StoryDAO() {
         this.id = -1;
-
     }
 
-    public StoryDAO(int id) {
+    public StoryDAO(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
